Odd behaviours partially fixed + readme
Forgot to add in speed to flee, now fixed making them flee at a sensible
speed instead of proportional to the distance.

Wander still correcting too much velocity wise. Maybe need to get them
to stop for a bit and then wander?

Added readme.md . not sure what to put in here as there's an absolute
bare minimum of stuff in it.